scanner-main/backend/requirements.txt
fastapi
uvicorn
pydantic
python-nmap
